Add HierarchicalIntegrityVerificationStorage

+++ src/LibHac/FsSystem/BitmapUtils.cs
@@ -5,11 +5,11 @@ namespace LibHac.FsSystem;
 public static class BitmapUtils
 {
     // ReSharper disable once InconsistentNaming
-    public static uint ILog2(uint value)
+    public static int ILog2(uint value)
     {
         Assert.SdkRequiresGreater(value, 0u);
 
-        const uint intBitCount = 32;
-        return intBitCount - 1 - (uint)Util.BitUtil.CountLeadingZeros(value);
+        const int intBitCount = 32;
+        return intBitCount - 1 - Util.BitUtil.CountLeadingZeros(value);
     }
 }
